完成合同服务项目的修改和解决了批量导入的问题，以及增加了代码注释

+++ src/service/employee/EmployeeService.java
@@ -121,7 +121,7 @@ public class EmployeeService {
         if(!result.success){
            return result;
         }
-        long[] eids = (long[]) result.extra;
+        int[] eids = (int[]) result.extra;
         for(int i = 0;i<eids.length;i++){//员工补充信息添加对应eid
             extras.get(i).setEid(eids[i]);
             payCards.get(i).setEid(eids[i]);
